Move GTO stuff into mrchem under utils
