+++ src/CMakeLists.txt
@@ -1,8 +1,6 @@
 link_directories(${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/lib)
 link_directories(${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR})
 
-add_subdirectory(initial_guess)
-
 add_library(mrchem STATIC
     SCFDriver.cpp
     parallel.cpp
@@ -16,24 +14,24 @@ add_subdirectory(analyticfunctions)
 add_subdirectory(qmfunctions)
 add_subdirectory(qmoperators)
 add_subdirectory(scf_solver)
+add_subdirectory(initial_guess)
 add_subdirectory(utils)
 target_include_directories(mrchem PRIVATE
     ${CMAKE_BINARY_DIR}
     ${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}
     ${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}/MRCPP
-    ${CMAKE_SOURCE_DIR}/src/chemistry
-    ${CMAKE_SOURCE_DIR}/src/properties
     ${CMAKE_SOURCE_DIR}/src/analyticfunctions
+    ${CMAKE_SOURCE_DIR}/src/chemistry
+    ${CMAKE_SOURCE_DIR}/src/initial_guess
+    ${CMAKE_SOURCE_DIR}/src/properties
     ${CMAKE_SOURCE_DIR}/src/qmfunctions
     ${CMAKE_SOURCE_DIR}/src/qmoperators
     ${CMAKE_SOURCE_DIR}/src/qmoperators/one_electron
     ${CMAKE_SOURCE_DIR}/src/qmoperators/two_electron
     ${CMAKE_SOURCE_DIR}/src/scf_solver
-    ${CMAKE_SOURCE_DIR}/src/initial_guess
+    ${CMAKE_SOURCE_DIR}/src/utils
     )
 target_link_libraries(mrchem
-    hydrogenguess
-    gtoguess
     mrcpp
     xcfun
     )
@@ -50,19 +48,18 @@ target_include_directories(mrchem.x PRIVATE
     ${CMAKE_BINARY_DIR}
     ${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}
     ${CMAKE_BINARY_DIR}/stage/${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}/MRCPP
-    ${CMAKE_SOURCE_DIR}/src/chemistry
-    ${CMAKE_SOURCE_DIR}/src/properties
     ${CMAKE_SOURCE_DIR}/src/analyticfunctions
+    ${CMAKE_SOURCE_DIR}/src/chemistry
+    ${CMAKE_SOURCE_DIR}/src/initial_guess
+    ${CMAKE_SOURCE_DIR}/src/properties
     ${CMAKE_SOURCE_DIR}/src/qmfunctions
     ${CMAKE_SOURCE_DIR}/src/qmoperators
     ${CMAKE_SOURCE_DIR}/src/qmoperators/one_electron
     ${CMAKE_SOURCE_DIR}/src/qmoperators/two_electron
     ${CMAKE_SOURCE_DIR}/src/scf_solver
-    ${CMAKE_SOURCE_DIR}/src/initial_guess
+    ${CMAKE_SOURCE_DIR}/src/utils
     )
 target_link_libraries(mrchem.x
-    hydrogenguess
-    gtoguess
     mrchem
     mrcpp
     getkw
