+++ src/utils/gto_utils/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(mrchem PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/AOBasis.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/AOContraction.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Intgrl.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/OrbitalExp.cpp
+)
